+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -139,26 +139,63 @@ resource "aws_ecs_service" "test_client" {
 module "test_client" {
   source = "../../../../../../modules/mesh-task"
   family = "test_client_${var.suffix}"
-  container_definitions = [{
-    name      = "basic"
-    image     = "docker.mirror.hashicorp.services/nicholasjackson/fake-service:v0.21.0"
-    essential = true
-    environment = [
-      {
-        name  = "UPSTREAM_URIS"
-        value = "http://localhost:1234"
+  container_definitions = [
+    {
+      name             = "basic"
+      image            = "docker.mirror.hashicorp.services/nicholasjackson/fake-service:v0.21.0"
+      essential        = true
+      logConfiguration = local.test_client_log_configuration
+      environment = [
+        {
+          name  = "UPSTREAM_URIS"
+          value = "http://localhost:1234"
+        }
+      ]
+      linuxParameters = {
+        initProcessEnabled = true
       }
-    ]
-    linuxParameters = {
-      initProcessEnabled = true
+      # Keep the client running for 10 seconds to validate graceful shutdown behavior.
+      entryPoint = ["/bin/sh", "-c", <<EOT
+/app/fake-service &
+export PID=$!
+trap "{ echo 'TEST LOG: on exit'; kill $PID; }" 0
+function onterm() {
+    echo "TEST LOG: Caught sigterm. Sleeping 10s..."
+    sleep 10
+    exit 0
+}
+trap onterm TERM
+wait $PID
+EOT
+      ]
+      healthCheck = {
+        command  = ["CMD-SHELL", "echo 1"]
+        interval = 30
+        retries  = 3
+        timeout  = 5
+      }
+    },
+    {
+      # Inject an additional container to monitor apps during task shutdown.
+      name             = "shutdown-monitor"
+      image            = "docker.mirror.hashicorp.services/golang:1.17-alpine"
+      essential        = false
+      logConfiguration = local.test_client_log_configuration
+      # AWS: "We do not enforce a size limit on the environment variables..."
+      # Then, the max environment var length is ~32k
+      environment = [{
+        name  = "GOLANG_MAIN_B64"
+        value = base64encode(file("${path.module}/shutdown-monitor.go"))
+      }]
+      # NOTE: `go run <file>` signal handling is different: https://github.com/golang/go/issues/40467
+      entryPoint = ["/bin/sh", "-c", <<EOT
+echo "$GOLANG_MAIN_B64" | base64 -d > main.go
+go build main.go
+exec ./main
+EOT
+      ]
     }
-    healthCheck = {
-      command  = ["CMD-SHELL", "echo 1"]
-      interval = 30
-      retries  = 3
-      timeout  = 5
-    }
-  }]
+  ]
   retry_join = module.consul_server.server_dns
   upstreams = [
     {
@@ -166,15 +203,8 @@ module "test_client" {
       local_bind_port  = 1234
     }
   ]
-  log_configuration = {
-    logDriver = "awslogs"
-    options = {
-      awslogs-group         = var.log_group_name
-      awslogs-region        = var.region
-      awslogs-stream-prefix = "test_client_${var.suffix}"
-    }
-  }
-  outbound_only = true
+  log_configuration = local.test_client_log_configuration
+  outbound_only     = true
 
   tls                            = var.secure
   consul_server_ca_cert_arn      = module.consul_server.ca_cert_arn
@@ -204,19 +234,13 @@ module "test_server" {
   source = "../../../../../../modules/mesh-task"
   family = "test_server_${var.suffix}"
   container_definitions = [{
-    name      = "basic"
-    image     = "docker.mirror.hashicorp.services/nicholasjackson/fake-service:v0.21.0"
-    essential = true
+    name             = "basic"
+    image            = "docker.mirror.hashicorp.services/nicholasjackson/fake-service:v0.21.0"
+    essential        = true
+    logConfiguration = local.test_server_log_configuration
   }]
-  retry_join = module.consul_server.server_dns
-  log_configuration = {
-    logDriver = "awslogs"
-    options = {
-      awslogs-group         = var.log_group_name
-      awslogs-region        = var.region
-      awslogs-stream-prefix = "test_server_${var.suffix}"
-    }
-  }
+  retry_join        = module.consul_server.server_dns
+  log_configuration = local.test_server_log_configuration
   checks = [
     {
       checkid  = "server-http"
@@ -236,4 +260,24 @@ module "test_server" {
   consul_client_token_secret_arn = var.secure ? module.acl_controller[0].client_token_secret_arn : ""
   acl_secret_name_prefix         = var.suffix
   consul_ecs_image               = var.consul_ecs_image
+}
+
+locals {
+  test_server_log_configuration = {
+    logDriver = "awslogs"
+    options = {
+      awslogs-group         = var.log_group_name
+      awslogs-region        = var.region
+      awslogs-stream-prefix = "test_server_${var.suffix}"
+    }
+  }
+
+  test_client_log_configuration = {
+    logDriver = "awslogs"
+    options = {
+      awslogs-group         = var.log_group_name
+      awslogs-region        = var.region
+      awslogs-stream-prefix = "test_client_${var.suffix}"
+    }
+  }
 }